makefiles: had a 'U' option to the ar call that was confusing the compiler - I removed this.
logflag.f: had a bug in that mistook temperature for a cloud - fixed.

ch4ch4_s.f: first line of 'BLOCK DATA INPUT' was confusing compiler as there is already a common block elsewhere called 'input' and this was being overwritten. changed to 'BLOCK DATA INPUTCH4CH4' so it is distinct.

Nick T